+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Accessibility improvements
 - Add multiple severity levels to sitewide alert
 - Fix: StreamField typo causing RecursionError in Wagtail 2.13
+- Upgrade Wagtail to version 2.13
 - Use the newly released django-gov-notify library for GOV.UK Notify integration
 
 ## 36.1 (2021-05-25)